pyproject.toml
[project]
name = "multi-agent-bac-runtime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bedrock-agentcore>=0.1.7",
    "boto3>=1.40.50",
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.0",
    "strands-agents>=1.12.0",
    "strands-agents-tools>=0.2.11",
    "uvicorn[standard]>=0.37.0",
]
